+++ fbrip/decoder_test.go
